backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
python-jose[cryptography]==3.3.0
pydantic==2.5.3
httpx==0.26.0
numpy==1.24.3
scipy==1.11.4
textblob==0.17.1
python-dotenv==1.0.0
